+++ python/GafferSceneTest/__init__.py
@@ -164,6 +164,7 @@ from .OptionTweaksTest import OptionTweaksTest
 from .OptionQueryTest import OptionQueryTest
 from .RenameTest import RenameTest
 from .PrimitiveVariableQueryTest import PrimitiveVariableQueryTest
+from .SetQueryTest import SetQueryTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *