+++ Changes.md
@@ -1,6 +1,11 @@
 1.1.x.x (relative to 1.1.5.0)
 =======
 
+Features
+--------
+
+- SetQuery : Added a new node for querying the set memberships of a location.
+
 Improvements
 ------------